Add documentation for countyBoundaries
